Pull all that's common to the Triggerable interface
